CA-20878: writing core dumps on SIGSEGV
To allow vncterm to write core dumps in the chroot directory we
use the same strategy as qemu:

when vncterm starts, it forks: the child chroot's and drops privileges,
then executes the rest of the code. The parent keeps the privileges and
waits for a message from the child or for a SIGCHLD.
If the parents receives a "sigsegv" message, it chown's the chroot
directory to allow the child to write the core dump.
Before exiting the parent removes the chroot directory, if it is empty.