Simplify privsep operations
We don't need to ask to write the statefile using privsep, because the
privileged parent process knows when the child has to write the
statefile: it receives a SIGUSR1 from xapi.
Hence we can remove all the "file descriptor through sockets" functions
and keep only one privsep operation: the request to write to xenstore
the path of the statefile.