Add qemu copyright info to debian/copyright
This copyright is tricky because the original vncterm authors copied a bunch of
code directly from qemu. This isn't a problem, except for writing the copyright
info.  qemu is licensed under "LGPL2, GPL2, and whatever else each file says it
is," which makes it difficult to describe in a Debian copyright file. I have
solved this by incorporating text from the qemu debian package's copyright
file, which explains the situation better than I could.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>