Add copyright/license information for individual qemu files
As per a conversation with Ian Campbell, we've modified the copyright file so
that it referes to individual qemu-based files by name, and lists their
original authors. This information came from the license headers of the
individual files. Where there was no license header, the file was grouped with
similar files, based on what was included by what.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>